rename a file in the folder contants
