+++ src/constants/carousel.ts
